make product and cart btn into components
