+++ src/features/pos/components/create-product-btn.tsx
@@ -1,0 +1,11 @@
+import { Button } from "@/components/ui/button";
+import { Plus } from 'lucide-react';
+
+export function CreateProductButton() {
+  return (
+    <Button className='bg-secondary hover:bg-emerald-500 px-4 rounded-[5px] font-semibold text-background'>
+      <span><Plus className='mr-1 font-bold' size={16} /></span>
+    Create Product
+    </Button>
+  );
+}
